+++ src/app/community/[symbol]/PostThumbnail.tsx
@@ -5,7 +5,7 @@ import { useRouter } from 'next/navigation';
 import { PostThumbnail as TPostThumbnail } from '@/types/post';
 import PostThumbnailCard from '@/components/PostThumbnailCard';
 import { useEffect, useState } from 'react';
-import { getCookie } from '@/util/cookie';
+import { getCookie } from '@/utils/cookie';
 import { TokenPayload } from '@/types/auth';
 import { TOKEN_KEY } from '@/constants/token';
 import { decodeJwt } from 'jose';
