+++ src/components/Comment/Comment.tsx
@@ -5,7 +5,7 @@ import { PrimaryDropdown, PrimaryDropdownProps } from '@/components/Dropdown';
 import Image from 'next/image';
 import { Comment as TComment } from '@/types/post';
 import DateFormatter from '../Dateformatter';
-import { getCookie } from '@/util/cookie';
+import { getCookie } from '@/utils/cookie';
 import { TOKEN_KEY } from '@/constants/token';
 import { decodeJwt } from 'jose';
 import { TokenPayload } from '@/types/auth';
