+++ src/utils/nickname-generator.ts
