+++ src/utils/optimizeImage.ts
