+++ src/utils/typeChecker.ts
